chore: rename lib/gui/style_utils.h => lib/gui/StyleUtils.h
