+++ src/apps/deskflow-gui/MainWindow.cpp
@@ -21,9 +21,9 @@
 #include "gui/Logger.h"
 #include "gui/Messages.h"
 #include "gui/StringUtils.h"
+#include "gui/StyleUtils.h"
 #include "gui/core/CoreProcess.h"
 #include "gui/ipc/DaemonIpcClient.h"
-#include "gui/style_utils.h"
 #include "gui/styles.h"
 #include "net/FingerprintDatabase.h"
 #include "platform/wayland.h"
