+++ src/lib/gui/StyleUtils.h
